Transaction Filter is added

+++ BMS/src/main/java/com/BMS/DTO/TransactionViewDtoV2.java
@@ -3,7 +3,7 @@ package com.BMS.DTO;
 import java.util.List;
 
 public record TransactionViewDtoV2(
-        int totalRecords,
+        long totalRecords,
         int totalPages,
         List<TransactionViewDto> data
 ) {
